requirements.txt
tensorflow-gpu==2.9
tensorflow-addons